Improve skills access and path display
